Move vrf code to their own files
